Added BestWithMods score status and fixed groupby in lb query
